% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{color_palettes}
\alias{color_palettes}
\title{Custom color palettes}
\format{list of different color palettes}
\usage{
data(color_palettes)
}
\description{
A dataset containing custom color palettes.
}
