% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_barplot.R
\name{custom_barplot}
\alias{custom_barplot}
\title{Plot customized barplot}
\usage{
custom_barplot(data, filename = NULL, x, fill_variable,
  custom_palette = NULL, title = "Plot title",
  ylab = "Fraction of patients")
}
\arguments{
\item{data}{matrix; contains variables x and fill_variable}

\item{filename}{character or \code{NULL}; filename for export or no export}

\item{x}{character; variable for x aesthetic}

\item{fill_variable}{character; variable for fill aesthetic}

\item{custom_palette}{character; define palette of fill_variable}

\item{title}{character}

\item{ylab}{character; y-axis label}
}
\description{
Creates a customized barchart and optionally exports it as a pdf.
}
