% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_boxplot.R
\name{custom_boxplot}
\alias{custom_boxplot}
\title{Plot customized boxplot}
\usage{
custom_boxplot(data, filename = NULL, width = 1.4, height = 1.55, x,
  y, ylab = NULL, log_y = FALSE, title = NULL, jitter = FALSE,
  custom_palette = NULL)
}
\arguments{
\item{data}{matrix; contains variables x and y}

\item{filename}{character or \code{NULL}; filename for export or no export}

\item{width}{numeric; width of pdf export in cm}

\item{height}{numeric; height of pdf export in cm}

\item{x}{character; variable for x aesthetic}

\item{y}{character; variable for y aesthetic}

\item{ylab}{character; y-axis label}

\item{log_y}{logical; switch logarithmic display of y-axis}

\item{title}{character; title of plot}

\item{jitter}{logical; overplot with jittered individual datapoints}

\item{custom_palette}{character; define color palette for x}
}
\description{
Creates a customized barchart and optionally exports it as a pdf.
}
