% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_FindMarkers.R
\name{custom_FindMarkers}
\alias{custom_FindMarkers}
\title{Perform differential gene expression analysis for select feature and select
cell subset; perform p-adjustment for number of subsets to test for;
this function is based on Seurat's FindMarkers function and adds custom variables to
the resulting data frame}
\usage{
custom_FindMarkers(
  data,
  idents = NULL,
  features,
  grouping_var = "genotype",
  ident.1 = NULL,
  min.pct = 0,
  logfc.threshold = 0
)
}
\arguments{
\item{data}{Seurat object}

\item{idents}{character; idents in Seurat object to perform DEA on;
passed to Seurat::FindMarkers}

\item{features}{character; genes in Seurat object to perform DEA on;
passed to Seurat::FindMarkers}

\item{grouping_var}{character; meta-feature of data to group for DEA;
passed to Seurat::FindMarkers}

\item{ident.1}{character; specify group (one level in grouping_var) for denominator);
passed to Seurat::FindMarkers}

\item{min.pct}{numeric; define minimum fraction of cells in idents to perform DEA on;
passed to Seurat::FindMarkers}

\item{logfc.threshold}{numeric; passed to Seurat::FindMarkers}
}
\description{
Takes a Seurat object as input and requires specification of
idents and features to test for
}
\author{
Benjamin N. Ostendorf
}
