% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{custom_linewidth}
\alias{custom_linewidth}
\title{Custom linewidth}
\format{
numeric value
}
\usage{
data(custom_linewidth)
}
\description{
A variable containing a custom linewidth equalling XX pt.
}
