% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_PCA.R
\name{custom_PCA}
\alias{custom_PCA}
\title{Plot customized PCA}
\usage{
custom_PCA(object, intgroup = "condition", ntop = 500, pointsize = 2,
  returnData = FALSE, custom_pal = NULL)
}
\arguments{
\item{object}{Deseq dataset object}

\item{intgroup}{character; variable defining groups}

\item{ntop}{numeric; number of genes with highes variance to include in analysis}

\item{returnData}{logical; return data rather than plot}

\item{custom_pal}{character; colors to customize group colors}
}
\description{
This function is customized version of DESeq2's plotPCA function.
}
