% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_qqplot.R
\name{custom_qqplot}
\alias{custom_qqplot}
\title{Customized Q-Q plot}
\usage{
custom_qqplot(data, var, group, width = NULL, height = NULL,
  xlab = NULL, ylab = NULL, no.labels = FALSE, pal = NULL,
  filename = NULL, title = TRUE, raster = TRUE,
  within.plot.label = FALSE, return.data = FALSE)
}
\arguments{
\item{data}{data.frame with columns with normalized gene expression values
and one column with a group stratifier, replicates/cells in rows}

\item{var}{character; name of gene to plot QQ-plot for, column in data}

\item{group}{character; column name data to use as stratifying variable in QQ-plot}

\item{width}{numeric; width of pdf export in cm}

\item{height}{numeric; height of pdf export in cm}

\item{pal}{character; custom palette for dot colours}

\item{filename}{character or \code{NULL}; filename for export or no export}

\item{return.data}{logical; return plot as list instead of printing}
}
\description{
Creates a customized QQ-Plot and optionally exports it as a pdf.
}
\author{
Benjamin N. Ostendorf
}
