% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_sig_levels.R
\name{custom_sig_levels}
\alias{custom_sig_levels}
\title{Add significance labels (*, **, ***, ****) to data frame}
\usage{
custom_sig_levels(data, column = "p_val")
}
\arguments{
\item{data}{data frame including one column named "p_val"}

\item{column}{character column holding p-values in data}
}
\description{
Takes a data frame as input with a column named `p_val` and returns the dataframe
with the additional column `signifcance` with significance levels
}
\author{
Benjamin N. Ostendorf
}
