% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_survplot.R
\name{custom_survplot}
\alias{custom_survplot}
\title{Plot customized survival graphs}
\usage{
custom_survplot(survFit, filename = NULL, title = NULL,
  xlab = "Overall survival (years)", ylab = "Survival probability",
  custom_palette = NULL, custom_legends = NULL, width = 2,
  height = 2.2)
}
\arguments{
\item{survFit}{survfit object}

\item{filename}{character or \code{NULL}; filename for export or no export}

\item{title}{character; plot title}

\item{xlab}{character; x-axis label}

\item{ylab}{character; y-axis label}

\item{custom_palette}{character; color palette for groups}

\item{custom_legends}{character; overwrite legend labels (see warning below)}

\item{width}{numeric; width of pdf export in cm}

\item{height}{numeric; height of pdf export in cm}
}
\description{
Creates a customized survival plot and optionally exports it as a pdf.
Partly inspired by the IMvigor2010CoreBiologies R package (Mariathasan, Nature, 2018).
}
